Re-organised repository
Added more of the initial And Gate tests and also re-organised the repository by putting files in folders
