+++ Random/data.txt
@@ -1,0 +1,1000 @@
+63R
+32B
+63R
+12B
+44R
+43R
+53R
+32B
+44R
+31B
+63R
+43R
+31B
+12B
+53R
+54R
+12B
+63R
+22B
+32B
+32B
+63R
+64R
+54R
+31B
+22B
+53R
+43R
+54R
+21B
+63R
+64R
+11B
+43R
+31B
+43R
+43R
+43R
+63R
+53R
+43R
+54R
+31B
+43R
+53R
+12B
+31B
+32B
+22B
+12B
+22B
+53R
+32B
+32B
+53R
+11B
+11B
+32B
+63R
+44R
+64R
+21B
+43R
+32B
+22B
+54R
+32B
+63R
+44R
+32B
+53R
+22B
+44R
+43R
+32B
+53R
+43R
+22B
+21B
+43R
+44R
+43R
+43R
+64R
+54R
+12B
+44R
+11B
+22B
+31B
+11B
+44R
+12B
+32B
+54R
+44R
+43R
+44R
+21B
+31B
+44R
+21B
+64R
+53R
+63R
+22B
+12B
+63R
+54R
+22B
+64R
+21B
+22B
+43R
+12B
+54R
+32B
+11B
+43R
+64R
+31B
+11B
+21B
+11B
+43R
+21B
+63R
+63R
+53R
+22B
+44R
+22B
+12B
+43R
+44R
+11B
+11B
+21B
+12B
+64R
+54R
+32B
+21B
+12B
+31B
+32B
+64R
+32B
+44R
+43R
+11B
+21B
+44R
+11B
+22B
+64R
+22B
+44R
+43R
+64R
+32B
+44R
+63R
+31B
+32B
+53R
+43R
+11B
+22B
+64R
+63R
+11B
+31B
+31B
+11B
+22B
+31B
+21B
+44R
+31B
+11B
+54R
+53R
+64R
+43R
+12B
+21B
+53R
+22B
+12B
+22B
+43R
+11B
+32B
+53R
+44R
+32B
+32B
+12B
+64R
+31B
+53R
+44R
+32B
+31B
+22B
+54R
+43R
+43R
+54R
+53R
+63R
+22B
+44R
+44R
+63R
+32B
+43R
+64R
+54R
+63R
+31B
+21B
+63R
+54R
+11B
+31B
+63R
+64R
+63R
+21B
+54R
+64R
+53R
+53R
+31B
+53R
+53R
+21B
+44R
+31B
+31B
+22B
+32B
+32B
+22B
+21B
+21B
+21B
+64R
+12B
+31B
+54R
+63R
+12B
+31B
+53R
+22B
+21B
+21B
+22B
+54R
+64R
+31B
+53R
+44R
+53R
+53R
+22B
+31B
+32B
+22B
+11B
+54R
+22B
+21B
+43R
+21B
+43R
+64R
+32B
+64R
+31B
+32B
+22B
+31B
+53R
+54R
+44R
+63R
+31B
+32B
+21B
+44R
+11B
+64R
+63R
+11B
+63R
+21B
+22B
+22B
+32B
+43R
+22B
+21B
+32B
+53R
+43R
+44R
+53R
+22B
+43R
+11B
+21B
+31B
+53R
+22B
+44R
+63R
+53R
+54R
+54R
+32B
+43R
+31B
+64R
+11B
+11B
+12B
+31B
+31B
+43R
+11B
+43R
+32B
+53R
+64R
+12B
+53R
+12B
+11B
+22B
+63R
+43R
+64R
+21B
+12B
+11B
+22B
+22B
+11B
+21B
+64R
+12B
+32B
+11B
+21B
+64R
+21B
+22B
+53R
+64R
+32B
+21B
+43R
+12B
+11B
+53R
+21B
+21B
+54R
+53R
+54R
+21B
+63R
+44R
+31B
+12B
+21B
+22B
+63R
+44R
+22B
+31B
+44R
+64R
+64R
+22B
+43R
+63R
+21B
+53R
+31B
+44R
+22B
+32B
+32B
+44R
+44R
+53R
+63R
+12B
+53R
+54R
+44R
+43R
+53R
+54R
+63R
+54R
+31B
+53R
+22B
+12B
+22B
+44R
+11B
+32B
+22B
+53R
+64R
+12B
+11B
+54R
+22B
+32B
+11B
+64R
+21B
+44R
+43R
+53R
+44R
+21B
+11B
+54R
+21B
+44R
+43R
+44R
+44R
+44R
+64R
+63R
+44R
+11B
+54R
+43R
+22B
+43R
+21B
+21B
+64R
+21B
+12B
+12B
+12B
+43R
+54R
+53R
+21B
+44R
+43R
+22B
+44R
+32B
+12B
+31B
+12B
+53R
+22B
+21B
+53R
+64R
+43R
+32B
+21B
+53R
+32B
+64R
+12B
+64R
+54R
+32B
+31B
+44R
+43R
+43R
+32B
+44R
+63R
+53R
+32B
+11B
+21B
+12B
+53R
+64R
+12B
+43R
+63R
+63R
+31B
+63R
+44R
+43R
+64R
+31B
+12B
+53R
+31B
+12B
+44R
+53R
+12B
+11B
+44R
+43R
+53R
+32B
+53R
+31B
+22B
+44R
+54R
+43R
+63R
+63R
+63R
+44R
+53R
+54R
+53R
+54R
+22B
+31B
+64R
+22B
+53R
+53R
+64R
+54R
+12B
+43R
+12B
+43R
+53R
+54R
+53R
+11B
+63R
+64R
+53R
+53R
+53R
+63R
+64R
+64R
+53R
+53R
+22B
+21B
+21B
+12B
+32B
+63R
+32B
+11B
+11B
+44R
+63R
+12B
+22B
+63R
+32B
+22B
+11B
+21B
+54R
+12B
+54R
+11B
+64R
+64R
+11B
+32B
+64R
+64R
+11B
+21B
+12B
+31B
+12B
+12B
+63R
+43R
+21B
+54R
+22B
+43R
+43R
+11B
+12B
+54R
+12B
+54R
+12B
+11B
+12B
+54R
+11B
+64R
+64R
+21B
+11B
+64R
+64R
+44R
+43R
+21B
+21B
+12B
+44R
+22B
+53R
+12B
+12B
+64R
+11B
+63R
+21B
+11B
+53R
+22B
+43R
+43R
+21B
+22B
+22B
+21B
+21B
+64R
+63R
+43R
+63R
+31B
+53R
+43R
+31B
+22B
+32B
+63R
+21B
+54R
+64R
+11B
+44R
+63R
+64R
+31B
+11B
+64R
+32B
+31B
+54R
+21B
+63R
+54R
+43R
+54R
+63R
+53R
+12B
+53R
+12B
+64R
+12B
+43R
+53R
+12B
+31B
+22B
+44R
+44R
+44R
+32B
+43R
+43R
+44R
+21B
+32B
+11B
+44R
+53R
+32B
+44R
+22B
+43R
+53R
+22B
+32B
+44R
+22B
+54R
+53R
+11B
+53R
+63R
+63R
+11B
+64R
+43R
+63R
+44R
+22B
+32B
+44R
+12B
+44R
+63R
+44R
+22B
+64R
+43R
+32B
+21B
+64R
+43R
+63R
+53R
+11B
+11B
+22B
+54R
+53R
+11B
+22B
+12B
+54R
+31B
+12B
+54R
+44R
+54R
+32B
+21B
+43R
+22B
+64R
+21B
+11B
+54R
+43R
+44R
+44R
+21B
+21B
+43R
+22B
+44R
+12B
+53R
+31B
+32B
+43R
+31B
+11B
+32B
+22B
+11B
+54R
+11B
+53R
+31B
+64R
+53R
+43R
+64R
+21B
+43R
+53R
+53R
+64R
+11B
+11B
+54R
+12B
+12B
+44R
+22B
+31B
+11B
+53R
+22B
+64R
+53R
+31B
+53R
+11B
+63R
+22B
+12B
+53R
+32B
+12B
+64R
+53R
+64R
+64R
+53R
+21B
+43R
+11B
+22B
+64R
+22B
+32B
+43R
+22B
+44R
+53R
+21B
+21B
+21B
+12B
+53R
+12B
+11B
+22B
+21B
+31B
+22B
+63R
+12B
+21B
+32B
+31B
+22B
+44R
+22B
+53R
+64R
+31B
+63R
+12B
+63R
+44R
+32B
+32B
+64R
+43R
+32B
+11B
+54R
+21B
+32B
+43R
+11B
+22B
+12B
+32B
+54R
+64R
+64R
+31B
+32B
+11B
+64R
+22B
+12B
+32B
+63R
+63R
+53R
+44R
+64R
+44R
+63R
+21B
+44R
+44R
+21B
+53R
+53R
+12B
+31B
+11B
+54R
+22B
+21B
+43R
+22B
+44R
+43R
+21B
+32B
+21B
+54R
+12B
+44R
+31B
+43R
+43R
+54R
+64R
+11B
+63R
+22B
+53R
+32B
+32B
+64R
+32B
+53R
+32B
+12B
+31B
+43R
+31B
+44R
+64R
+22B
+54R
+64R
+53R
+31B
+64R
+64R
+43R
+43R
+53R
+54R
+53R
+32B
+63R
+44R
+53R
+53R
+53R
+11B
+63R
+11B
+21B
+31B
+44R
+12B
+63R
+44R
+54R
+64R
+11B
+53R
+54R
+63R
+64R
+44R
+11B
+12B
+44R
+22B
+63R
+21B
+12B
+63R
+44R
+21B
+12B
+22B
+12B
+44R
+21B
+22B
+63R
+64R
+32B
+53R
+32B
+53R
+22B
+54R
+53R
+32B
+21B
+31B
+44R
+12B
+63R
+12B
+12B
+32B
+54R
+44R
+54R
+44R
+44R
+63R
+43R
+64R
+64R
